refactor(i18n): recursively re-export nested resources in /admin-form
