+++ frontend/src/i18n/locales/features/admin-form/en-sg.ts
@@ -1,0 +1,7 @@
+import { enSG as navbar } from './navbar'
+import { enSG as sidebar } from './sidebar'
+
+export const enSG = {
+    navbar,
+    sidebar,
+}
