+++ frontend/src/i18n/locales/features/admin-form/index.ts
@@ -1,0 +1,6 @@
+export { type Navbar } from './navbar'
+export { type Fields } from './sidebar'
+export { type HeaderAndInstructions } from './sidebar'
+export { type Logic } from './sidebar'
+export { type ThankYou } from './sidebar'
+export * from './en-sg'
