+++ frontend/src/i18n/locales/features/admin-form/sidebar/en-sg.ts
@@ -1,0 +1,11 @@
+import { enSG as fields } from './fields'
+import { enSG as headerAndInstructions } from './header-and-instructions'
+import { enSG as logic } from './logic'
+import { enSG as thankYou } from './thank-you'
+
+export const enSG = {
+    fields,
+    headerAndInstructions,
+    logic,
+    thankYou,
+}
