+++ frontend/src/i18n/locales/features/admin-form/sidebar/index.ts
@@ -1,0 +1,5 @@
+export { type Fields } from './fields'
+export { type HeaderAndInstructions } from './header-and-instructions'
+export { type Logic } from './logic'
+export { type ThankYou } from './thank-you'
+export * from './en-sg'
