+++ frontend/src/i18n/locales/features/en-sg.ts
@@ -1,0 +1,13 @@
+import { enSG as adminForm } from './admin-form'
+import { enSG as common } from './common'
+import { enSG as login } from './login'
+import { enSG as publicForm } from './public-form'
+
+export const enSG = {
+    features: {
+        adminForm,
+        common,
+        login,
+        publicForm,
+    }
+}
