+++ frontend/src/i18n/locales/features/index.ts
@@ -1,8 +1,9 @@
-export { type Navbar } from './admin-form/navbar'
-export { type Fields } from './admin-form/sidebar/fields'
-export { type HeaderAndInstructions } from './admin-form/sidebar/header-and-instructions'
-export { type Logic } from './admin-form/sidebar/logic'
-export { type ThankYou } from './admin-form/sidebar/thank-you'
+export { type Navbar } from './admin-form'
+export { type Fields } from './admin-form'
+export { type HeaderAndInstructions } from './admin-form'
+export { type Logic } from './admin-form'
+export { type ThankYou } from './admin-form'
 export { type Common } from './common'
 export { type Login } from './login'
 export { type PublicForm } from './public-form'
+export * from './en-sg'
