+++ frontend/src/i18n/locales/types.ts
@@ -3,21 +3,38 @@ import { Navbar, Common, Login, PublicForm, Fields, HeaderAndInstructions, Logic
 interface Translation {
   translation: {
     features: {
-      navbar?: Navbar
+      adminForm: {
+        sidebar: {
+          fields?: Fields
+          headerAndInstructions?: HeaderAndInstructions
+          logic?: Logic
+          thankYou?: ThankYou
+        }
+        navbar?: Navbar
+      }
       common?: Common
       publicForm?: PublicForm
       login?: Login
-      fields?: Fields
-      headerAndInstructions?: HeaderAndInstructions
-      logic?: Logic
-      thankYou?: ThankYou
     }
   }
 }
 
 export interface FallbackTranslation extends Translation {
   translation: {
-    features: Required<Translation['translation']['features']>
+    features: {
+      adminForm: {
+        sidebar: {
+          fields: Fields
+          headerAndInstructions: HeaderAndInstructions
+          logic: Logic
+          thankYou: ThankYou
+        }
+        navbar: Navbar
+      }
+      common: Common
+      publicForm: PublicForm
+      login: Login
+    }
   }
 }
